Move CancellationReasonSerializer
